ISSUE#2974: extract shared JiraClient into scripts/cve/jira_client.py
Both create_cve_trackers.py and cve_due_dates.py had duplicate JiraClient
class implementations. Extract into a shared module with the superset of
methods from both copies.
