+++ scripts/cve/cve_due_dates.py
@@ -29,32 +29,13 @@ Requires:
 """
 
 import argparse
-import json
-import os
 import re
 import sys
 from dataclasses import dataclass, field
 from datetime import datetime, date
 
-from scripts.cve import create_ssl_context
-from scripts.cve.jira_auth import (
-    JiraAuthError,
-    get_auth_headers,
-    get_cached_api_base_url,
-    resolve_cloud_base_url,
-)
-
-try:
-    import requests
-    HAS_REQUESTS = True
-except ImportError:
-    import urllib.request
-    import urllib.error
-    import urllib.parse
-    HAS_REQUESTS = False
-
-
-_SSL_CONTEXT = create_ssl_context() if not HAS_REQUESTS else None
+from scripts.cve.jira_auth import JiraAuthError
+from scripts.cve.jira_client import JiraClient
 
 
 @dataclass
@@ -86,114 +67,6 @@ class TrackerInfo:
         """True if tracker has no due date but linked issues do."""
         return self.due_date is None and self.earliest_child_due_date is not None
 
-
-JIRA_DEFAULT_URL = "https://redhat.atlassian.net"
-
-
-class JiraClient:
-    """Simple Jira REST API v3 client."""
-
-    def __init__(self, base_url: str, auth_headers: dict | None = None):
-        """Direct constructor — testable, no env var dependencies."""
-        self.base_url = base_url.rstrip("/")
-        self.headers = {"Content-Type": "application/json"}
-        if auth_headers:
-            self.headers.update(auth_headers)
-
-    @classmethod
-    def from_env(cls) -> "JiraClient":
-        """Factory that reads env vars, resolves auth and base URL.
-
-        For OAuth tokens the Jira REST API must be accessed through the
-        Atlassian API gateway (``api.atlassian.com/ex/jira/{cloudId}``).
-        For API-token (Basic) or legacy Bearer auth the configured
-        ``JIRA_URL`` is used directly.
-        """
-        jira_url = os.environ.get("JIRA_URL", JIRA_DEFAULT_URL)
-        auth_headers = get_auth_headers(jira_url)
-
-        base_url = jira_url
-        auth_value = auth_headers.get("Authorization", "")
-
-        if auth_value.startswith("Bearer ") and not os.environ.get("JIRA_TOKEN", "").strip():
-            cached_base = get_cached_api_base_url(jira_url)
-            if cached_base:
-                base_url = cached_base
-            else:
-                token = auth_value.removeprefix("Bearer ")
-                base_url = resolve_cloud_base_url(token, jira_url)
-
-        return cls(base_url, auth_headers)
-
-    def _request(self, method: str, endpoint: str, params: dict | None = None, data: dict | None = None) -> dict:
-        """Make a request to the Jira API."""
-        url = f"{self.base_url}{endpoint}"
-
-        if HAS_REQUESTS:
-            response = requests.request(
-                method,
-                url,
-                params=params,
-                json=data,
-                headers=self.headers,
-                timeout=30
-            )
-            response.raise_for_status()
-            if response.text:
-                return response.json()
-            return {}
-        else:
-            if params:
-                query_string = "&".join(f"{k}={urllib.parse.quote(str(v))}" for k, v in params.items())
-                url = f"{url}?{query_string}"
-
-            req = urllib.request.Request(url, headers=self.headers, method=method)
-            if data:
-                req.data = json.dumps(data).encode("utf-8")
-
-            with urllib.request.urlopen(req, context=_SSL_CONTEXT) as resp:
-                content = resp.read().decode()
-                if content:
-                    return json.loads(content)
-                return {}
-
-    def search_issues(self, jql: str, fields: str = "key,summary,status,labels,duedate,issuelinks",
-                      max_results: int = 500) -> list[dict]:
-        """Search for issues using JQL (API v3, token-based pagination)."""
-        all_issues: list[dict] = []
-        next_page_token: str | None = None
-
-        while len(all_issues) < max_results:
-            params: dict[str, str | int] = {
-                "jql": jql,
-                "maxResults": min(100, max_results - len(all_issues)),
-                "fields": fields,
-            }
-            if next_page_token:
-                params["nextPageToken"] = next_page_token
-
-            data = self._request("GET", "/rest/api/3/search/jql", params=params)
-            issues = data.get("issues", [])
-            all_issues.extend(issues)
-
-            if data.get("isLast", True) or not issues:
-                break
-
-            next_page_token = data.get("nextPageToken")
-            if not next_page_token:
-                break
-
-        return all_issues
-
-    def get_issue(self, issue_key: str, fields: str = "key,summary,status,duedate,issuelinks") -> dict:
-        """Get a single issue."""
-        params = {"fields": fields}
-        return self._request("GET", f"/rest/api/3/issue/{issue_key}", params=params)
-
-    def update_issue(self, issue_key: str, fields: dict) -> None:
-        """Update an issue's fields."""
-        data = {"fields": fields}
-        self._request("PUT", f"/rest/api/3/issue/{issue_key}", data=data)
 
 
 def extract_cve_id(text: str) -> str | None:
@@ -229,7 +102,7 @@ def find_cve_trackers(client: JiraClient, max_results: int = 500) -> list[Tracke
 
     # Get RHAIENG CVE issues (trackers)
     jql = 'project = RHAIENG AND labels in ("CVE") AND resolution = unresolved ORDER BY duedate ASC'
-    issues = client.search_issues(jql, max_results=max_results)
+    issues = client.search_issues(jql, fields="key,summary,status,labels,duedate,issuelinks", max_results=max_results)
     print(f"Found {len(issues)} RHAIENG CVE tracker issues")
 
     trackers = []
